# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Kompas engine API
  version: "1.0.0"
  description: |
    Scoring engine for Kompas (Kraków + Praha). Owner: Workstream B.
    Contract v1 (tag contracts-v1). Changes after v1 are ADDITIVE ONLY and are listed in docs/notes/B.md.

    Conventions
    - All labels that come from config are I18nText objects ({pl, cs, en, ko}) so a language switch needs no refetch.
    - Server-rendered sentences (explanations, relax hint) are plain strings in the requested `lang`.
    - Scores are integers 0–100. Coordinates are WGS84 (EPSG:4326). H3 cells are res 9 unless stated.
    - Archetypes are returned as ids; the web app owns their names ("archetype.<id>").
    - Errors use the Error schema with HTTP 400/404/422/503.
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/{city}/meta:
    get:
      tags: [core]
      operationId: getMeta
      parameters:
        - $ref: "#/components/parameters/City"
        - $ref: "#/components/parameters/Lang"
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/Meta"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/{city}/grid:
    get:
      tags: [core]
      operationId: getGrid
      description: "The city's grid.geojson (H3 res-9 polygons; properties h3, district_id, district_name, neighborhood, habitable, population_est; numeric feature id = row index). Static, cache it. Same file as data/processed/{city}/grid.geojson."
      parameters:
        - $ref: "#/components/parameters/City"
      responses:
        "200":
          description: GeoJSON FeatureCollection
          content: {application/geo+json: {schema: {type: object}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/{city}/score:
    post:
      tags: [core]
      operationId: score
      parameters:
        - $ref: "#/components/parameters/City"
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ScoreRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/ScoreResponse"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/Invalid"}
  /api/{city}/place/{h3}:
    get:
      tags: [core]
      operationId: getPlace
      parameters:
        - $ref: "#/components/parameters/City"
        - {name: h3, in: path, required: true, schema: {$ref: "#/components/schemas/H3"}}
        - $ref: "#/components/parameters/Lang"
        - name: state
          in: query
          required: false
          description: |
            base64url (no padding) of the JSON of a ScoreRequest (persona, weights, anchors, filters, budget).
            Without it the place is scored with the persona "custom" defaults and no anchors.
          schema: {type: string}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/Place"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/{city}/commute:
    get:
      tags: [core]
      operationId: getCommute
      description: Travel minutes from every habitable res-9 cell to the res-8 cell that contains (lat, lon). Unreachable (>120 min) cells are omitted.
      parameters:
        - $ref: "#/components/parameters/City"
        - {name: lat, in: query, required: true, schema: {type: number}}
        - {name: lon, in: query, required: true, schema: {type: number}}
        - {name: mode, in: query, required: false, schema: {$ref: "#/components/schemas/Mode"}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/CommuteResponse"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/Invalid"}
  /api/{city}/districts:
    get:
      tags: [core]
      operationId: listDistricts
      description: Admin units of the city (static info). For a ranking use POST /score with aggregate=district.
      parameters:
        - $ref: "#/components/parameters/City"
        - {name: level, in: query, required: false, schema: {type: string, enum: [district, neighborhood], default: district}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/DistrictList"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/{city}/similar/{h3}:
    get:
      tags: [ml]
      operationId: getSimilar
      parameters:
        - $ref: "#/components/parameters/City"
        - {name: h3, in: path, required: true, schema: {$ref: "#/components/schemas/H3"}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 20, default: 5}}
        - $ref: "#/components/parameters/Lang"
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/SimilarResponse"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/twins:
    get:
      tags: [ml]
      operationId: getTwins
      description: "Twin neighbourhoods across cities (distance similarity on the pooled cross-city feature vector: 1 − d / median random-pair distance, clipped to 0–1). Give either h3 or district."
      parameters:
        - {name: from, in: query, required: true, schema: {$ref: "#/components/schemas/CityId"}}
        - {name: to, in: query, required: true, schema: {$ref: "#/components/schemas/CityId"}}
        - {name: h3, in: query, required: false, schema: {$ref: "#/components/schemas/H3"}}
        - {name: district, in: query, required: false, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 20, default: 5}}
        - $ref: "#/components/parameters/Lang"
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/TwinsResponse"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/Invalid"}
  /api/{city}/geocode:
    get:
      tags: [core]
      operationId: geocode
      description: Local, accent-insensitive address/POI/place search ("zizkov" → "Žižkov"); Photon fallback.
      parameters:
        - $ref: "#/components/parameters/City"
        - {name: q, in: query, required: true, schema: {type: string, minLength: 2}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 20, default: 8}}
        - $ref: "#/components/parameters/Lang"
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/GeocodeResponse"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /api/{city}/prefs/pairs:
    get:
      tags: [ml]
      operationId: getPrefPairs
      description: (P2) Pairs of contrasting place profiles for preference learning.
      parameters:
        - $ref: "#/components/parameters/City"
        - {name: n, in: query, required: false, schema: {type: integer, minimum: 2, maximum: 12, default: 6}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/PrefPairs"}}}}
  /api/prefs/fit:
    post:
      tags: [ml]
      operationId: fitPrefs
      description: (P2) Bradley–Terry logistic regression on criterion-difference vectors → suggested emoji levels.
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/PrefFitRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/PrefFitResponse"}}}}
        "422": {$ref: "#/components/responses/Invalid"}
components:
  parameters:
    City:
      name: city
      in: path
      required: true
      schema: {$ref: "#/components/schemas/CityId"}
    Lang:
      name: lang
      in: query
      required: false
      description: Language of server-rendered sentences. Defaults to the city's defaultLang.
      schema: {$ref: "#/components/schemas/Lang"}
  responses:
    NotFound:
      description: Unknown city, cell or district
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    Invalid:
      description: Invalid request
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
...
    CityId: {type: string, enum: [krakow, praha]}
...
    Mode: {type: string, enum: [transit, bike, walk], default: transit}
...
    H3: {type: string, pattern: "^[0-9a-f]{15}$", examples: ["891e2040d4bffff"]}
...
    Meta:
      type: object
      required: [city, levels, criteria, personas, mustHaveCategories, archetypes, modes, price, adminLevels, manifest]
      properties:
        city: {$ref: "#/components/schemas/CityId"}
        dataVersion: {type: string}
        dataSource: {$ref: "#/components/schemas/DataSource"}
        walkSpeedKmh: {type: number}
        levels:
          type: array
          items:
            type: object
            required: [level, weight, emoji, label]
            properties:
              level: {$ref: "#/components/schemas/Level"}
              weight: {type: number}
              emoji: {type: string}
              label: {$ref: "#/components/schemas/I18nText"}
        criteria: {type: array, items: {$ref: "#/components/schemas/CriterionMeta"}}
        personas: {type: array, items: {$ref: "#/components/schemas/Persona"}}
        mustHaveCategories:
          type: array
          items:
            type: object
            required: [id, emoji, label, available]
            properties:
              id: {type: string}
              emoji: {type: string}
              label: {$ref: "#/components/schemas/I18nText"}
              available: {type: boolean, description: false if the category does not exist in this city or has no data}
        archetypes:
          type: array
          items:
            type: object
            required: [id, emoji]
            properties: {id: {$ref: "#/components/schemas/ArchetypeId"}, emoji: {type: string}}
        modes:
          type: array
          items:
            type: object
            required: [id, emoji, label]
            properties: {id: {$ref: "#/components/schemas/Mode"}, emoji: {type: string}, label: {$ref: "#/components/schemas/I18nText"}}
        price:
          type: object
          required: [indicator, mode, currency, unit, filterField, budgetField]
          properties:
            indicator: {type: string, examples: [buy_per_m2, rent_per_m2]}
            mode: {type: string, enum: [buy, rent]}
            currency: {$ref: "#/components/schemas/Currency"}
            unit: {$ref: "#/components/schemas/I18nText"}
            filterField: {type: string, enum: [maxPricePerM2, maxRentPerM2]}
            budgetField: {type: string, enum: [total, monthlyRent]}
            cityMedian: {type: [number, "null"]}
            min: {type: [number, "null"]}
            max: {type: [number, "null"]}
        adminLevels:
          type: object
          required: [district, neighborhood]
          properties:
            district: {$ref: "#/components/schemas/AdminLevel"}
            neighborhood: {$ref: "#/components/schemas/AdminLevel"}
        manifest: {type: array, items: {$ref: "#/components/schemas/ManifestEntry"}}
...
    ScoreRequest:
      type: object
      required: [weights]
      properties:
        v: {type: integer, const: 1, default: 1}
        lang: {$ref: "#/components/schemas/Lang"}
        persona: {type: [string, "null"], description: "persona id; drives indicatorWeights and walkFactor"}
        weights: {$ref: "#/components/schemas/CriterionLevels"}
        indicatorWeights: {type: object, additionalProperties: {type: number}, description: "optional column → weight override on top of the persona"}
        anchors: {type: array, maxItems: 3, items: {$ref: "#/components/schemas/Anchor"}, default: []}
        filters: {$ref: "#/components/schemas/Filters"}
        budget: {$ref: "#/components/schemas/Budget"}
        aggregate: {type: string, enum: [hex, district, neighborhood], default: hex}
        limit: {type: integer, minimum: 1, maximum: 100, default: 20}
        includeCells: {type: boolean, default: true, description: "false → cells: [] (smaller payload)"}
        rankBy: {type: [string, "null"], description: "what `top` is ordered by: null = match %, a CriterionId = that criterion's score, \"anchor:<id>\" = shortest travel time to that anchor; unknown values rank by match %. `score` stays the match %"}
        district: {type: [string, "null"], description: "aggregate=hex only: restrict `top` to the cells of this district id"}
...
    ScoreResponse:
      type: object
      required: [city, currency, lang, computeMs, aggregate, count, cells, top, relaxHint, excludedCriteria, weightsUsed, cityMedian]
      properties:
        city: {$ref: "#/components/schemas/CityId"}
        currency: {$ref: "#/components/schemas/Currency"}
        lang: {$ref: "#/components/schemas/Lang"}
        computeMs: {type: number}
        aggregate: {type: string, enum: [hex, district, neighborhood]}
        count:
          type: object
          required: [cells, habitable, passing]
          properties: {cells: {type: integer}, habitable: {type: integer}, passing: {type: integer}}
        cells: {type: array, items: {$ref: "#/components/schemas/CellTuple"}}
        top: {type: array, items: {$ref: "#/components/schemas/RankedPlace"}}
        relaxHint: {oneOf: [{$ref: "#/components/schemas/RelaxHint"}, {type: "null"}]}
        excludedCriteria: {type: array, items: {$ref: "#/components/schemas/CriterionId"}, description: "level>0 but no data in this city (or commute without anchors)"}
        weightsUsed: {$ref: "#/components/schemas/CriterionLevels"}
        cityMedian: {$ref: "#/components/schemas/CriterionScores"}

    # ───────────────────────── place
...
    Place:
      type: object
      required: [id, kind, city, centroid, name, district, habitable, criteria, cityMedian, indicators, nearest, anchors]
      properties:
        id: {$ref: "#/components/schemas/H3"}
        kind: {type: string, enum: [hex]}
        city: {$ref: "#/components/schemas/CityId"}
        lang: {$ref: "#/components/schemas/Lang"}
        centroid: {$ref: "#/components/schemas/LatLon"}
        name: {type: string}
        neighborhood: {type: [string, "null"]}
        district: {oneOf: [{$ref: "#/components/schemas/PlaceRef"}, {type: "null"}]}
        habitable: {type: boolean}
        populationEst: {type: [integer, "null"]}
        score: {type: [integer, "null"]}
        passes: {type: [boolean, "null"]}
        criteria: {$ref: "#/components/schemas/CriterionScores"}
        cityMedian: {$ref: "#/components/schemas/CriterionScores"}
        indicators: {type: array, items: {$ref: "#/components/schemas/IndicatorValue"}}
        nearest: {type: array, items: {$ref: "#/components/schemas/NearestPoi"}}
        anchors: {type: array, items: {$ref: "#/components/schemas/AnchorTime"}}
        price: {oneOf: [{$ref: "#/components/schemas/PriceInfo"}, {type: "null"}]}
        budgetM2: {type: [number, "null"]}
        budgetText: {type: [string, "null"]}
        archetype: {oneOf: [{$ref: "#/components/schemas/ArchetypeDetail"}, {type: "null"}]}
        highlights: {type: array, items: {$ref: "#/components/schemas/Explanation"}}
        warnings: {type: array, items: {$ref: "#/components/schemas/Explanation"}}

    # ───────────────────────── commute / districts
    CommuteResponse:
      type: object
      required: [city, anchorCell, mode, maxMinutes, cells]
      properties:
        city: {$ref: "#/components/schemas/CityId"}
        anchorCell: {type: string, description: res-8 H3 cell used as the destination}
        mode: {$ref: "#/components/schemas/Mode"}
        maxMinutes: {type: integer, const: 120}
        cells:
          type: array
          description: "[res-9 h3, minutes]"
          items:
            type: array
            prefixItems: [{$ref: "#/components/schemas/H3"}, {type: integer, minimum: 0, maximum: 120}]
            minItems: 2
            maxItems: 2
        computeMs: {type: number}
    DistrictList:
      type: object
      required: [city, level, items]
      properties:
        city: {$ref: "#/components/schemas/CityId"}
        level: {type: string, enum: [district, neighborhood]}
        items:
          type: array
          items:
            type: object
            required: [id, name, centroid, cellCount]
            properties:
              id: {type: string}
              name: {type: string}
              centroid: {$ref: "#/components/schemas/LatLon"}
              cellCount: {type: integer}
              habitableCount: {type: integer}
              populationEst: {type: [integer, "null"]}

    # ───────────────────────── ML
...
    SimilarResponse:
      type: object
      required: [city, from, items]
      properties:
        city: {$ref: "#/components/schemas/CityId"}
        from: {$ref: "#/components/schemas/H3"}
        items: {type: array, items: {$ref: "#/components/schemas/SimilarPlace"}}
    TwinsResponse:
      type: object
      required: [from, to, items]
      properties:
        from:
          type: object
          required: [city, id, kind, name, centroid]
          properties:
            city: {$ref: "#/components/schemas/CityId"}
            id: {type: string}
            kind: {type: string, enum: [hex, district]}
            name: {type: string}
            district: {oneOf: [{$ref: "#/components/schemas/PlaceRef"}, {type: "null"}]}
            centroid: {$ref: "#/components/schemas/LatLon"}
            archetype: {oneOf: [{$ref: "#/components/schemas/ArchetypeRef"}, {type: "null"}]}
        to: {$ref: "#/components/schemas/CityId"}
        items: {type: array, items: {$ref: "#/components/schemas/SimilarPlace"}}
    PrefPairs:
      type: object
      required: [city, pairs]
      properties:
        city: {$ref: "#/components/schemas/CityId"}
        pairs:
          type: array
          items:
            type: object
            required: [id, a, b]
            properties:
              id: {type: string}
              a: {$ref: "#/components/schemas/CriterionScores"}
              b: {$ref: "#/components/schemas/CriterionScores"}
    PrefFitRequest:
      type: object
      required: [choices]
      properties:
        choices:
          type: array
          minItems: 1
          items:
            type: object
            required: [a, b, choice]
            properties:
              a: {$ref: "#/components/schemas/CriterionScores"}
              b: {$ref: "#/components/schemas/CriterionScores"}
              choice: {type: string, enum: [a, b]}
...
    GeocodeResponse:
      type: object
      required: [query, items]
      properties:
        query: {type: string}
        items:
          type: array
          items:
            type: object
            required: [label, kind, lat, lon, source]
            properties:
              label: {type: string}
              sublabel: {type: [string, "null"]}
              kind: {type: string, enum: [address, poi, place, district]}
              category: {type: [string, "null"], description: POI category when kind=poi (e.g. university)}
              lat: {type: number}
              lon: {type: number}
              h3: {type: [string, "null"]}
              source: {type: string, enum: [local, photon]}
...